JPA repository, dependency injection, database seeding

+++ src/main/java/com/apiweb/entities/User.java
@@ -21,7 +21,7 @@ public class User implements Serializable{
 	private static final long serialVersionUID = 1L;
 	
 	@Id @GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	
 	private String name;
 	private String email;
@@ -33,7 +33,7 @@ public class User implements Serializable{
 	}
 
 
-	public User(int id, String name, String email, String phone, String password) {
+	public User(Long id, String name, String email, String phone, String password) {
 		super();
 		this.id = id;
 		this.name = name;
